+++ pkgbuilder/upgrade.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-# PKGBUILDer v4.2.2
+# PKGBUILDer v4.2.3
 # An AUR helper (and library) in Python 3.
 # Copyright © 2011-2015, Chris Warrick.
 # See /LICENSE for licensing information.
@@ -201,9 +201,14 @@ def auto_upgrade(downgrade=False, vcsup=False, fetchonly=False):
         else:
             query = ':: ' + _('Proceed with installation? [Y/n] ')
 
-        yesno = input(query)
+        if DS.confirm:
+            yesno = input(query)
 
-        if yesno.lower().strip().startswith('y') or yesno.strip() == '':
-            return upgnames
+            if yesno.lower().strip().startswith('y') or yesno.strip() == '':
+                return upgnames
+            else:
+                return []
         else:
-            return []
+            # Print the query and then return immediately, pacman does that too
+            print(query)
+            return upgnames